Use Users Command as the Root of Users Ops Command
This moves more command to work in both auth and users command.
This is to prevent breaking the cli for anyone using it as we transition
to removing the auth command completly.

While auth command will still work, by v1 it is going to be completly
removed.

Replace Auth with Users To Make Commands Shorter #21